Enable adapting resolution rules for Spark 2.x (previously Spark 3.x only)
