+++ hudi-client/hudi-spark-client/src/main/scala/org/apache/spark/sql/hudi/SparkAdapter.scala
@@ -116,12 +116,7 @@ trait SparkAdapter extends Serializable {
   def getFilePartitions(sparkSession: SparkSession, partitionedFiles: Seq[PartitionedFile],
       maxSplitBytes: Long): Seq[FilePartition]
 
-  def resolveHoodieTable(plan: LogicalPlan): Option[CatalogTable] = {
-    EliminateSubqueryAliases(plan) match {
-      case LogicalRelation(_, _, Some(table), _) if isHoodieTable(table) => Some(table)
-      case _ => None
-    }
-  }
+  def resolveHoodieTable(plan: LogicalPlan): Option[CatalogTable]
 
   def isHoodieTable(map: java.util.Map[String, String]): Boolean = {
     map.getOrDefault("provider", "").equals("hudi")
