+++ hudi-spark-datasource/hudi-spark3.3.x/src/main/scala/org/apache/spark/sql/adapter/Spark3_3Adapter.scala
@@ -58,7 +58,7 @@ class Spark3_3Adapter extends BaseSpark3Adapter {
 
   override def resolveHoodieTable(plan: LogicalPlan): Option[CatalogTable] = {
     EliminateSubqueryAliases(plan) match {
-      case HoodieLogicalRelation(LogicalRelation(_, _, Some(table), _)) => Some(table)
+      case HoodieLogicalRelationAdapter(LogicalRelation(_, _, Some(table), _)) => Some(table)
       case LogicalRelation(_, _, Some(table), _) if isHoodieTable(table) => Some(table)
       case DataSourceV2Relation(v2Table: V2TableWithV1Fallback, _, _, _, _) if isHoodieTable(v2Table.v1Table) => Some(v2Table.v1Table)
       case _ => None
